+++ lib/Slave/Slave.h
@@ -112,7 +112,7 @@ public:
     void sendBackLeftRPM(uint16_t rpm);
     void sendBackRightRPM(uint16_t rpm);
     void sendGoals(uint16_t yellowAngle, uint16_t yellowDistance, uint16_t blueAngle, uint16_t blueDistance);
-    void sendRobotPosition(uint16_t x, uint16_t y);
+    void sendRobotPosition(int8_t x, int8_t y);
     DebugSettings getDebugSettings();
 
     void updateDebugSettings();
